+++ lading_payload/src/common.rs
@@ -1,6 +1,5 @@
 //! Common utilities for all lading payloads
 
 pub mod config;
-pub(crate) mod interner;
 pub(crate) mod strings;
 pub(crate) mod tags;